third_party/blink/web_tests/http/tests/loading/pdf-commit-load-callbacks-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - BeginNavigation request to 'data:application/pdf;base64,', http method GET
frame "<!--framePath //<!--frame0-->-->" - DidStartNavigation
main frame - didFinishDocumentLoadForFrame
frame "<!--framePath //<!--frame0-->-->" - didFailProvisionalLoadWithError
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
